Refactor DedicatedAccountClient to use pydantic types
