+++ src/pypaystack2/paystack.py
@@ -3,9 +3,8 @@ from pypaystack2.sub_clients.async_clients.apple_pay import AsyncApplePayClient
 from pypaystack2.sub_clients.async_clients.bulk_charges import AsyncBulkChargeClient
 from pypaystack2.sub_clients.async_clients.charge import AsyncChargeClient
 from pypaystack2.sub_clients.async_clients.customers import AsyncCustomerClient
-from pypaystack2.sub_clients.dedicated_accounts import (
+from pypaystack2.sub_clients.async_clients.dedicated_accounts import (
     AsyncDedicatedAccountClient,
-    DedicatedAccountClient,
 )
 from pypaystack2.sub_clients.disputes import AsyncDisputeClient, DisputeClient
 from pypaystack2.sub_clients.integration import (
@@ -46,6 +45,9 @@ from pypaystack2.sub_clients.sync_clients.bulk_charges import (
 from pypaystack2.sub_clients.sync_clients.charge import ChargeClient
 from pypaystack2.sub_clients.sync_clients.customers import (
     CustomerClient,
+)
+from pypaystack2.sub_clients.sync_clients.dedicated_accounts import (
+    DedicatedAccountClient,
 )
 from pypaystack2.sub_clients.terminals import TerminalClient, AsyncTerminalClient
 from pypaystack2.sub_clients.transactions import (
