+++ tests/test_sub_clients/test_sync_clients/test_dedicated_accounts.py
@@ -4,26 +4,17 @@ import httpx
 from dotenv import load_dotenv
 
 from pypaystack2.sub_clients import DedicatedAccountClient
-from tests.test_sub_clients.mocked_api_testcase import MockedAPITestCase
-
-
-class MockedDedicatedAccountTestCase(MockedAPITestCase):
-    @classmethod
-    def setUpClass(cls) -> None:
-        super().setUpClass()
-        load_dotenv()
-        cls.wrapper = DedicatedAccountClient()
 
 
 class DedicatedAccountTestCase(TestCase):
     @classmethod
     def setUpClass(cls) -> None:
         load_dotenv()
-        cls.wrapper = DedicatedAccountClient()
+        cls.client = DedicatedAccountClient()
 
     def test_can_create(self):
         # TODO: Test properly.
-        self.wrapper.create(customer="CUS_5qmwswiljybyyne")
+        self.client.create(customer="CUS_5qmwswiljybyyne")
 
     def test_can_get_dedicated_accounts(self): ...
 
@@ -39,5 +30,5 @@ class DedicatedAccountTestCase(TestCase):
 
     def test_can_get_providers(self):
         # TODO: Test properly.
-        response = self.wrapper.get_providers()
+        response = self.client.get_providers()
         self.assertEqual(response.status_code, httpx.codes.UNAUTHORIZED)
